requirements.txt
pyyaml
xmltodict
pytest
